// src/app/pages/pkmn/interfaces/pkmns.ts
export interface Pkmns {
    count: number;
    next: string;
    previous: null;
    results: Result[];
}
export interface Result{
    name: string;
    url: string;
    data?: Pokemon;
}
export interface Pokemon {
    abilities:                Ability[];
    base_experience:          number;
    cries:                    Cries;
    forms:                    Species[];
    game_indices:             GameIndex[];
    height:                   number;
    held_items:               any[];
    id:                       number;
    is_default:               boolean;
    location_area_encounters: string;
    moves:                    Move[];
    name:                     string;
    order:                    number;
    past_abilities:           any[];
    past_types:               any[];
    species:                  Species;
    sprites:                  Sprites;
    stats:                    Stat[];
    types:                    Type[];
    weight:                   number;
}